+++ serialization/src/commonMain/kotlin/nl/adaptivity/xmlutil/serialization/XMLEncoder.kt
@@ -174,7 +174,7 @@ internal open class XmlEncoderBase internal constructor(
         ) {
             when  {
                 xmlDescriptor is XmlContextualDescriptor -> {
-                    val actualDescriptor = xmlDescriptor.resolve(serializer, config, serializersModule)
+                    val actualDescriptor = xmlDescriptor.resolve(serializer.descriptor, config, serializersModule)
                     val delegateEncoder = XmlEncoder(actualDescriptor, elementIndex, discriminatorName)
                     actualDescriptor.effectiveSerializationStrategy(serializer).serializeSafe(delegateEncoder, value)
                 }